+++ quota/mysqlqm/mysql_quota.go
@@ -79,27 +79,6 @@ func (m *QuotaManager) GetTokens(ctx context.Context, numTokens int, specs []quo
 	return nil
 }
 
-// PeekTokens implements quota.Manager.PeekTokens.
-// Global/Write tokens reflect the number of rows in the Unsequenced tables, other specs are
-// considered infinite.
-func (m *QuotaManager) PeekTokens(ctx context.Context, specs []quota.Spec) (map[quota.Spec]int, error) {
-	tokens := make(map[quota.Spec]int)
-	for _, spec := range specs {
-		var num int
-		if spec.Group == quota.Global && spec.Kind == quota.Write {
-			count, err := m.countUnsequenced(ctx)
-			if err != nil {
-				return nil, err
-			}
-			num = m.MaxUnsequencedRows - count
-		} else {
-			num = quota.MaxTokens
-		}
-		tokens[spec] = num
-	}
-	return tokens, nil
-}
-
 // PutTokens implements quota.Manager.PutTokens.
 // It's a noop for QuotaManager.
 func (m *QuotaManager) PutTokens(ctx context.Context, numTokens int, specs []quota.Spec) error {